% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{Assign}
\alias{Assign}
\title{Dose Assignment/Recommendation Given MTD Estiamte and Highest Assigned Dose, Avoid Skipping Dose}
\usage{
Assign(mtd, max.dose)
}
\arguments{
\item{mtd}{MTD estimate}

\item{max.dose}{highest assigned dose so far}
}
\description{
Dose Assignment/Recommendation Given MTD Estiamte and Highest Assigned Dose, Avoid Skipping Dose
}
