% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{CRM}
\alias{CRM}
\title{Continual Reassessment Method (CRM) with Multiple Toxicity Constraints}
\usage{
CRM(param.ini = NULL, dose, tox, skeleton, target,
  working.model = "power", cohort.size = 1,
  allow.dose.skipping = FALSE, allow.escalation.on.toxicity = FALSE)
}
\arguments{
\item{param.ini}{initial value of paramameter, default = rep(0, L)}

\item{dose}{dose level sequence}

\item{tox}{toxicity sequence}

\item{skeleton}{mapped skeleton for dose levels, a vector of length = # of dose levels (K), should be between 0 and 1}

\item{target}{targets for (cumulative) toxicity rates, a vector of length = # of toxicity levels (L)}

\item{working.model}{working model, character, default = "power" (currently only supports "power")}

\item{cohort.size}{default = 1}

\item{allow.dose.skipping}{default = FALSE}

\item{allow.escalation.on.toxicity}{default = FALSE}
}
\description{
Continual Reassessment Method (CRM) with Multiple Toxicity Constraints
}
