% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{CRMMLE}
\alias{CRMMLE}
\title{Continual Reassessment Method (CRM) by Maximum Likelihood Estimate (MLE)}
\usage{
CRMMLE(dose.tox, skeleton, model = "power", target, max.dose = NULL,
  ...)
}
\arguments{
\item{dose.tox}{counts of dose-toxicity, an integer matrix where nrow = # of dose levels, and ncol = (# of toxicity levels + 1)}

\item{skeleton}{mapped skeleton for dose levels}

\item{model}{working model, character, default = "power"}

\item{target}{target cumulative toxicity rates, a vector of length = # of toxicity levels}

\item{max.dose}{highest assigned dose so far}

\item{...}{other (fixed) parameters of working model}
}
\description{
Continual Reassessment Method (CRM) by Maximum Likelihood Estimate (MLE)
}
