% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{FitModel}
\alias{FitModel}
\title{Fit Working Model by Maximum Likelihood Estimate (MLE)}
\usage{
FitModel(param.ini = NULL, dose.tox, skeleton, working.model = "power")
}
\arguments{
\item{param.ini}{initial value of paramameter, default = rep(0, L)}

\item{dose.tox}{counts of dose-toxicity, an integer matrix where nrow = # of dose levels (K), and ncol = # of toxicity levels (L) + 1.}

\item{skeleton}{mapped skeleton for dose levels, a vector of length = # of dose levels (K), should be between 0 and 1}

\item{working.model}{working model, character, default = "power" (currently only supports "power")}
}
\description{
Fit Working Model by Maximum Likelihood Estimate (MLE)
}
