% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{FitWorkingModelMLE}
\alias{FitWorkingModelMLE}
\title{Fit Working Model by Maximum Likelihood Estimate (MLE)}
\usage{
FitWorkingModelMLE(dose.tox, skeleton, model = "power", ...)
}
\arguments{
\item{dose.tox}{counts of dose-toxicity, an integer matrix where nrow = # of dose levels, and ncol = (# of toxicity levels + 1)}

\item{skeleton}{mapped skeleton for dose levels}

\item{model}{working model, character, default = "power"}

\item{...}{other (fixed) parameters of working model}
}
\description{
Fit Working Model by Maximum Likelihood Estimate (MLE)
}
