% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{LogLik}
\alias{LogLik}
\title{Compute the Log Probability of Dose-Toxicity Data, Given Toxicity Rates}
\usage{
LogLik(dose.tox, tox.rate)
}
\arguments{
\item{dose.tox}{counts of dose-toxicity, an integer matrix where nrow = # of dose levels, and ncol = (# of toxicity levels + 1)}

\item{tox.rate}{toxicity rates computed from model parameter, a matrix of the same dimension as dose.tox, rowSums(tox.rate) should be ones}
}
\description{
Compute the Log Probability of Dose-Toxicity Data, Given Toxicity Rates
}
