% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{LogProb}
\alias{LogProb}
\title{Compute the Log Probability of Dose-Toxicity Data, Given Toxicity Rates}
\usage{
LogProb(dose.tox, tox.rate)
}
\arguments{
\item{dose.tox}{counts of dose-toxicity, an integer matrix where nrow = # of dose levels (K), and ncol = # of toxicity levels (L) + 1.}

\item{tox.rate}{(cumulative) toxicity rates computed from model parameter, a matrix where nrow = # of dose levels (K), and ncol = # of toxicity levels (L).}
}
\description{
Compute the Log Probability of Dose-Toxicity Data, Given Toxicity Rates
}
