% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{MTD}
\alias{MTD}
\title{Compute Maximum Tolerated Dose (MTD) from Toxicity Rates and Target}
\usage{
MTD(tox.rate, target)
}
\arguments{
\item{tox.rate}{(cumulative) toxicity rates computed from model parameter, a matrix where nrow = # of dose levels (K), and ncol = # of toxicity levels (L).}

\item{target}{targets for (cumulative) toxicity rates, a vector of length = # of toxicity levels (L)}
}
\description{
Compute Maximum Tolerated Dose (MTD) from Toxicity Rates and Target
}
