% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro.R
\name{PROCRM}
\alias{PROCRM}
\title{PRO-CRM wrapper}
\usage{
PROCRM(param.ini = NULL, dose, tox.c, tox.p, skeleton, target, method)
}
\arguments{
\item{param.ini}{initial value of paramameter, default = rep(0, L)}

\item{dose}{dose level sequence}

\item{tox.c}{toxicity outcome from clinitian}

\item{tox.p}{toxicity outcome from patient}

\item{skeleton}{mapped skeleton for dose levels, a vector of length = # of dose levels (K), should be between 0 and 1}

\item{target}{targets for (cumulative) toxicity rates, a vector of length = # of toxicity levels (L)}

\item{method}{one of "mar", "jntmar", "jntjnt"}
}
\description{
PRO-CRM wrapper
}
