% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro.R
\name{RTox}
\alias{RTox}
\title{Generate Patient-Reported Outcome (PRO), Given Toxicity Rates from Clinician, Patient, and Clinician/Patient}
\usage{
RTox(prob.c, prob.p, prob.cop)
}
\arguments{
\item{prob.c}{toxicity rate from clinician}

\item{prob.p}{toxicity rate from patient}

\item{prob.cop}{toxicity rate from clinician or patient}
}
\description{
Generate Patient-Reported Outcome (PRO), Given Toxicity Rates from Clinician, Patient, and Clinician/Patient
}
