% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{Seq2DoseTox}
\alias{Seq2DoseTox}
\title{Convert Trial Sequence to Dose-Toxicity Data (Counts)}
\usage{
Seq2DoseTox(dose, tox, n.dose, n.tox)
}
\arguments{
\item{dose}{dose level sequence}

\item{tox}{toxicity sequence}

\item{n.dose}{number of possible dose levels}

\item{n.tox}{number of possible toxicity levels (0 not included)}
}
\description{
Convert Trial Sequence to Dose-Toxicity Data (Counts)
}
