% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{SimOneCRM}
\alias{SimOneCRM}
\title{Single-Trial Simulation for CRM()}
\usage{
SimOneCRM(n, true.tox.rate, skeleton, target, working.model = "power",
  cohort.size = 1, allow.dose.skipping = FALSE,
  allow.escalation.on.toxicity = FALSE, seed = NULL)
}
\arguments{
\item{n}{# of subjects (sample size) in each trial, should be a multiple of cohort.size}

\item{true.tox.rate}{true (cumulative) toxicity rates, a real matrix where nrow = # of dose levels (K), and ncol = # of toxicity levels (L).}

\item{skeleton}{mapped skeleton for dose levels, a vector of length = # of dose levels (K), should be between 0 and 1}

\item{target}{targets for (cumulative) toxicity rates, a vector of length = # of toxicity levels (L)}

\item{working.model}{working model, character, default = "power" (currently only supports "power")}

\item{cohort.size}{default = 1}

\item{allow.dose.skipping}{default = FALSE}

\item{allow.escalation.on.toxicity}{default = FALSE}

\item{seed}{seed for RNG}
}
\description{
Single-Trial Simulation for CRM()
}
