% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{ToxRate}
\alias{ToxRate}
\title{Compute (Cumulative) Toxicity Rates for Working Model}
\usage{
ToxRate(param, skeleton, working.model = "power")
}
\arguments{
\item{param}{parameters of working model, a real vector of length = # of tocicity levels (L)}

\item{skeleton}{mapped skeleton for dose levels, a vector of length = # of dose levels (K), should be between 0 and 1}

\item{working.model}{working model, character, default = "power" (currently only supports "power")}
}
\description{
Compute (Cumulative) Toxicity Rates for Working Model
}
