% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{WorkingModel}
\alias{WorkingModel}
\title{Compute Toxicity Rates for Working Model,}
\usage{
WorkingModel(param, skeleton, model = "power", ...)
}
\arguments{
\item{param}{parameters of working model}

\item{skeleton}{mapped skeleton for dose levels}

\item{model}{working model, character, default = "power"}

\item{...}{other (fixed) parameters of working model}
}
\description{
Compute Toxicity Rates for Working Model,
}
