% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{WorkingModelPower}
\alias{WorkingModelPower}
\title{Compute Toxicity Rates for Power Working Model, Given Parameters and Skeleton}
\usage{
WorkingModelPower(b, skeleton)
}
\arguments{
\item{b}{parameters of power working model, a real vector of length = # of toxicity levels}

\item{skeleton}{mapped skeleton for dose levels, a vector of length = # of dose levels, should be between 0 and 1}
}
\description{
Compute Toxicity Rates for Power Working Model, Given Parameters and Skeleton
}
