% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm.R
\name{MTD}
\alias{MTD}
\title{Compute Maimum Tolerated Dose (MTD) from Working Model}
\usage{
MTD(param, skeleton, model = "power", target, ...)
}
\arguments{
\item{param}{parameters of working model, a real vector of length = # of toxicity levels}

\item{skeleton}{mapped skeleton for dose levels}

\item{model}{working model, character, default = "power"}

\item{target}{target cumulative toxicity rates, a vector of length = # of toxicity levels}

\item{...}{other (fixed) parameters of working model}
}
\description{
Compute Maimum Tolerated Dose (MTD) from Working Model
}
